+++ packages/perseus/src/index.ts
@@ -96,8 +96,6 @@ export {
 } from "./util/sizing-utils";
 export {mathOnlyParser} from "./widgets/interactive-graphs/utils";
 export {
-    getAnswersFromWidgets,
-    injectWidgets,
     isWrongAnswerSupported,
     shouldHaveIndividualAnswer,
     isWidgetIdInContent,
@@ -127,14 +125,15 @@ export {
      * @deprecated - import this function from perseus-core instead
      */
     isFailure,
+    getAnswersFromWidgets,
+    injectWidgets,
+    getPerseusAIData,
 } from "@khanacademy/perseus-core";
 
 export {
     generateTestRadioWidget,
     generateTestInteractiveGraphWidget,
     generateTestCategorizerWidget,
-    generateTestExpressionWidget,
-    generateTestNumericInputWidget,
 } from "./util/test-utils";
 export {
     getWidgetTypeByWidgetId,