+++ packages/perseus/src/widget-ai-utils/explanation/explanation-ai-utils.test.ts
@@ -1,36 +1,27 @@
+import {
+    generateTestPerseusRenderer,
+    type PerseusRenderer,
+    generateExplanationOptions,
+    generateExplanationWidget,
+} from "@khanacademy/perseus-core";
+
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
 import {getPromptJSON} from "./explanation-ai-utils";
 
-import type {PerseusRenderer} from "@khanacademy/perseus-core";
-
-const question1: PerseusRenderer = {
+const question1: PerseusRenderer = generateTestPerseusRenderer({
     content:
         "Here's the explanation\n[[\u2603 explanation 1]]\nDid you get that?",
-    images: {},
     widgets: {
-        "explanation 1": {
-            graded: true,
-            version: {
-                major: 0,
-                minor: 0,
-            },
-            // NOTE: The explanation widget doesn't consume this directly,
-            // instead, Perseus renders an overlay <div /> over top of the
-            // widget that intercepts interactions to it.
-            static: false,
-            type: "explanation",
-            options: {
+        "explanation 1": generateExplanationWidget({
+            options: generateExplanationOptions({
                 hidePrompt: "Hide explanation!",
-                widgets: {},
                 explanation: "This is an explanation",
-                static: false,
                 showPrompt: "Explanation",
-            },
-            alignment: "default",
-        },
+            }),
+        }),
     },
-};
+});
 
 describe("Explanation getPromptJSON", () => {
     it("it returns JSON with the expected format and fields", () => {